# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "restaurant api",
        "description": "Restaurants API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/restaurant": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get a list restaurants",
                "operationId": "7a110d3d9aaf2c008e33edbb3160007d",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "restaurants": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Prof. Albert Rolfson DDS"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Ipsam totam officia incidunt alias quam et quos. Quisquam culpa est ut. Repellendus accusantium quis quia doloribus."
                                                            },
                                                            "slug": {
                                                                "type": "string",
                                                                "example": "prof-albert-rolfson-dds"
                                                            },
                                                            "user": {
                                                                "type": "string",
                                                                "example": "admin"
                                                            },
                                                            "links": {
                                                                "properties": {
                                                                    "self": {
                                                                        "type": "string",
                                                                        "format": "uri",
                                                                        "example": "http://127.0.0.1:8000/api/v1/restaurant/prof-albert-rolfson-dds"
                                                                    },
                                                                    "index": {
                                                                        "type": "string",
                                                                        "format": "uri",
                                                                        "example": "http://127.0.0.1:8000/api/v1/restaurant"
                                                                    },
                                                                    "store": {
                                                                        "type": "string",
                                                                        "format": "uri",
                                                                        "example": "http://127.0.0.1:8000/api/v1/restaurant"
                                                                    },
                                                                    "update": {
                                                                        "type": "string",
                                                                        "format": "uri",
                                                                        "example": "http://127.0.0.1:8000/api/v1/restaurant/prof-albert-rolfson-dds"
                                                                    },
                                                                    "delete": {
                                                                        "type": "string",
                                                                        "format": "uri",
                                                                        "example": "http://127.0.0.1:8000/api/v1/restaurant/prof-albert-rolfson-dds"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Store a restaurant",
                "operationId": "3c46849c7ff8b8ea3de8bd311671d156",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Prof. Albert Rolfson DDS"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Ipsam totam officia incidunt alias quam et quos."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "restaurant": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Prof. Albert Rolfson DDS"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Ipsam totam officia incidunt alias quam et quos. Quisquam culpa est ut. Repellendus accusantium quis quia doloribus."
                                                        },
                                                        "slug": {
                                                            "type": "string",
                                                            "example": "prof-albert-rolfson-dds"
                                                        },
                                                        "user": {
                                                            "type": "string",
                                                            "example": "admin"
                                                        },
                                                        "links": {
                                                            "properties": {
                                                                "self": {
                                                                    "type": "string",
                                                                    "format": "uri",
                                                                    "example": "http://127.0.0.1:8000/api/v1/restaurant/prof-albert-rolfson-dds"
                                                                },
                                                                "index": {
                                                                    "type": "string",
                                                                    "format": "uri",
                                                                    "example": "http://127.0.0.1:8000/api/v1/restaurant"
                                                                },
                                                                "store": {
                                                                    "type": "string",
                                                                    "format": "uri",
                                                                    "example": "http://127.0.0.1:8000/api/v1/restaurant"
                                                                },
                                                                "update": {
                                                                    "type": "string",
                                                                    "format": "uri",
                                                                    "example": "http://127.0.0.1:8000/api/v1/restaurant/prof-albert-rolfson-dds"
                                                                },
                                                                "delete": {
                                                                    "type": "string",
                                                                    "format": "uri",
                                                                    "example": "http://127.0.0.1:8000/api/v1/restaurant/prof-albert-rolfson-dds"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/restaurant/{restaurant}": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get a restaurant",
                "operationId": "49ced671915de6760614454a0ec6c020",
                "parameters": [
                    {
                        "name": "restaurant",
                        "in": "path",
                        "description": "slug restaurant",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "name-slug"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show restaurant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "restaurant": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Prof. Albert Rolfson DDS"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Ipsam totam officia incidunt alias quam et quos. Quisquam culpa est ut."
                                                        },
                                                        "slug": {
                                                            "type": "string",
                                                            "example": "prof-albert-rolfson-dds"
                                                        },
                                                        "user": {
                                                            "type": "string",
                                                            "example": "admin"
                                                        },
                                                        "links": {
                                                            "properties": {
                                                                "self": {
                                                                    "type": "string",
                                                                    "format": "uri",
                                                                    "example": "http://127.0.0.1:8000/api/v1/restaurant/prof-albert-rolfson-dds"
                                                                },
                                                                "index": {
                                                                    "type": "string",
                                                                    "format": "uri",
                                                                    "example": "http://127.0.0.1:8000/api/v1/restaurant"
                                                                },
                                                                "store": {
                                                                    "type": "string",
                                                                    "format": "uri",
                                                                    "example": "http://127.0.0.1:8000/api/v1/restaurant"
                                                                },
                                                                "update": {
                                                                    "type": "string",
                                                                    "format": "uri",
                                                                    "example": "http://127.0.0.1:8000/api/v1/restaurant/prof-albert-rolfson-dds"
                                                                },
                                                                "delete": {
                                                                    "type": "string",
                                                                    "format": "uri",
                                                                    "example": "http://127.0.0.1:8000/api/v1/restaurant/prof-albert-rolfson-dds"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/restaurant/{id}": {
            "put": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Updated a restaurant",
                "operationId": "87c0c24c784026a3f3aefdfb13c9c1a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID the restaurant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Show restaurant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "restaurant": {
                                                    "$ref": "#/components/schemas/Restaurant"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Destroy a restaurant",
                "operationId": "31c2631caf5d1e75d97ac966983fcd9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID the restaurant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Restaurant": {
                "required": [
                    "name",
                    "description",
                    "slug",
                    "user_id"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the restaurant",
                        "type": "string",
                        "default": "My Restaurant"
                    },
                    "description": {
                        "description": "Description of the restaurant",
                        "type": "string",
                        "default": "My description"
                    },
                    "slug": {
                        "description": "Slug of the restaurant",
                        "type": "string",
                        "default": "my-restaurant"
                    },
                    "user_id": {
                        "description": "ID of the user",
                        "type": "integer",
                        "default": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer &lt;token&gt;)"
            }
        }
    },
...
}